[Android] ChromeActivity moved.
Moved BraveActivity to the same package.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/ac5ddcc168e18575d854c931420985265b7b8587

commit ac5ddcc168e18575d854c931420985265b7b8587
Author: Brandon Wylie <wylieb@chromium.org>
Date:   Thu Aug 20 03:57:14 2020 +0000

    Move ChromeActivity to the org.chromium.chrome.browser.app package

    internal counterpart: https://chrome-internal-review.googlesource.com/c/clank/internal/apps/+/3190243

    Bug: 1055811
