+++ android/java/org/chromium/chrome/browser/settings/BraveSyncScreensPreference.java
@@ -70,9 +70,9 @@ import org.chromium.base.Log;
 import org.chromium.base.ThreadUtils;
 import org.chromium.base.task.PostTask;
 import org.chromium.chrome.R;
-import org.chromium.chrome.browser.BraveActivity;
 import org.chromium.chrome.browser.BraveSyncWorker;
 import org.chromium.chrome.browser.BraveSyncReflectionUtils;
+import org.chromium.chrome.browser.app.BraveActivity;
 import org.chromium.chrome.browser.qrreader.BarcodeTracker;
 import org.chromium.chrome.browser.qrreader.BarcodeTrackerFactory;
 import org.chromium.chrome.browser.qrreader.CameraSource;
