+++ android/java/org/chromium/chrome/browser/toolbar/bottom/BraveBottomToolbarConfiguration.java
@@ -12,7 +12,7 @@ import android.view.Display;
 
 import org.chromium.base.ApplicationStatus;
 import org.chromium.base.ContextUtils;
-import org.chromium.chrome.browser.ChromeActivity;
+import org.chromium.chrome.browser.app.ChromeActivity;
 import org.chromium.chrome.browser.preferences.BravePreferenceKeys;
 import org.chromium.chrome.browser.preferences.SharedPreferencesManager;
