--- patches/chrome-android-java-src-org-chromium-chrome-browser-ChromeActivity.java.patch
@@ -1,13 +1,0 @@
-diff --git a/chrome/android/java/src/org/chromium/chrome/browser/ChromeActivity.java b/chrome/android/java/src/org/chromium/chrome/browser/ChromeActivity.java
-index 32dd5687aeb053a93892673db3facbd179ee4801..9f91999a6ab05eddcab8909987820c0d35696833 100644
---- a/chrome/android/java/src/org/chromium/chrome/browser/ChromeActivity.java
-+++ b/chrome/android/java/src/org/chromium/chrome/browser/ChromeActivity.java
-@@ -312,7 +312,7 @@ public abstract class ChromeActivity<C extends ChromeActivityComponent>
-     NightModeReparentingController mNightModeReparentingController;
- 
-     protected ChromeActivity() {
--        mIntentHandler = new IntentHandler(this, createIntentHandlerDelegate());
-+        mIntentHandler = new BraveIntentHandler(this, createIntentHandlerDelegate());
-     }
- 
-     @Override
