+++ patches/chrome-android-java-src-org-chromium-chrome-browser-app-ChromeActivity.java.patch
@@ -1,0 +1,13 @@
+diff --git a/chrome/android/java/src/org/chromium/chrome/browser/app/ChromeActivity.java b/chrome/android/java/src/org/chromium/chrome/browser/app/ChromeActivity.java
+index 4c33f321173e22c22c09905500c64e4553c75661..a40de28749ac5252c583309336b9faa2cda6d07f 100644
+--- a/chrome/android/java/src/org/chromium/chrome/browser/app/ChromeActivity.java
++++ b/chrome/android/java/src/org/chromium/chrome/browser/app/ChromeActivity.java
+@@ -323,7 +323,7 @@ public abstract class ChromeActivity<C extends ChromeActivityComponent>
+     NightModeReparentingController mNightModeReparentingController;
+ 
+     protected ChromeActivity() {
+-        mIntentHandler = new IntentHandler(this, createIntentHandlerDelegate());
++        mIntentHandler = new org.chromium.chrome.browser.BraveIntentHandler(this, createIntentHandlerDelegate());
+     }
+ 
+     @Override
